requirements.txt
spotipy
dbus-python
PyGObject
